Stop printing in keytableprinter (rename keytable)
